requirements.txt
attrs==22.1.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycryptodome==3.15.0
pyparsing==3.0.9
pytest==7.1.3
tomli==2.0.1
